+++ js/data.js
@@ -1,0 +1,30 @@
+//Массив возможных сообщений написаных под фото
+const USERS_MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+//имена людей оставляющих коментарии под фото
+const USERS_NAME = [
+  'Артём',
+  'Максим',
+  'Егор',
+  'Катя',
+  'Лиза',
+  'Даша',
+  'Дима'
+];
+
+//описание фотограии пользователя
+const DESCRIPT_IMAGE = [
+  'Отдыхаю на море',
+  'С пацанами в бане',
+  'Гуляю по ночному Уралмашу',
+  'Поздравляю всех с 3 сентября!!'
+];
+
+export{USERS_MESSAGE, USERS_NAME, DESCRIPT_IMAGE};
